% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tamp.R
\name{calc_tamp}
\alias{calc_tamp}
\title{Calculate the temperature amplitude from long term weather data}
\usage{
calc_tamp(wth)
}
\arguments{
\item{wth}{a data frame containing long term weather data including
columns of date of measurement (DATE), daily maximum temperature (TMAX),
and daily minimum temperature (TMIN)}
}
\description{
Calculate the temperature amplitude from long term weather data
}
