% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tav.R
\name{calc_tav}
\alias{calc_tav}
\title{Calculate the temperature average from long term weather data}
\usage{
calc_tav(wth)
}
\arguments{
\item{wth}{a data frame containing long term weather data including
columns of date of measurement (DATE), daily maximum temperature (TMAX),
and daily minimum temperature (TMIN)}
}
\description{
Calculate the temperature average from long term weather data
}
