% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{f_to_c}
\alias{f_to_c}
\title{Convert degrees Fahrenheit to degrees Celsius}
\usage{
f_to_c(t_f)
}
\arguments{
\item{t_f}{a vector of temperature in degrees Fahrenheit}
}
\value{
a vector of temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit to degrees Celsius
}
