% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_power.R
\name{get_nasa_power}
\alias{get_nasa_power}
\title{Get NASA POWER agroclimatology data for crop modeling}
\usage{
get_nasa_power(
  lat,
  long,
  start,
  end,
  pars = c("PRECTOTCORR", "T2M_MAX", "T2M_MIN", "ALLSKY_SFC_SW_DWN", "WS2M", "RH2M"),
  INSI = "NASA",
  ...
)
}
\arguments{
\item{lat}{latitude of query point in decimal degrees}

\item{long}{longitude of query point in decimal degrees}

\item{start}{first date for data query as POSIXt, Date, or
character formatted as four-digit year, two-digit month,
and two-digit day of month (e.g. "2010-12-31")}

\item{end}{last date for data query as POSIXt, Date, or
character formatted as four-digit year, two-digit month,
and two-digit day of month (e.g. 2010-12-31)}

\item{pars}{a character vector of parameters to download. A full
parameter dictionary can be found at:
\url{https://power.larc.nasa.gov/#resources}}

\item{INSI}{a four-digit character code to uniquely identify the
query location}

\item{...}{additional parameters passed to \code{\link[nasapower:get_power]{nasapower::get_power()}}}
}
\description{
Get NASA POWER agroclimatology data for crop modeling
}
