% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filling.R
\name{lin_interp}
\alias{lin_interp}
\title{Linearly interpolate data}
\usage{
lin_interp(time_series)
}
\arguments{
\item{time_series}{vector of data to be interpolated}
}
\description{
Linearly interpolate data
}
