% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversions.R
\name{m_s_to_km_d}
\alias{m_s_to_km_d}
\title{Convert wind speed from meters per second to kilometers per day}
\usage{
m_s_to_km_d(m_s)
}
\arguments{
\item{m_s}{a vector of wind speed in meters per second}
}
\value{
a vector of wind speed in kilometers per day
}
\description{
Convert wind speed from meters per second to kilometers per day
}
