% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wth_2m_wind.R
\name{wth_2m_wind}
\alias{wth_2m_wind}
\title{Adjust wind speed to 2 meter height}
\usage{
wth_2m_wind(wind, height)
}
\arguments{
\item{wind}{the wind speed measured at a given height}

\item{height}{the height of wind speed measurement in meters}
}
\description{
Adjust wind speed measured at an arbitrary height to wind speed at 2 meters
based on equation 47 from the FAO Irrigation and Drainage Paper 56 (Allen et al., 1998)
}
