% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wth_drop_units.R
\name{wth_drop_units}
\alias{wth_drop_units}
\title{Drop units from a weather data frame}
\usage{
wth_drop_units(wth_df)
}
\arguments{
\item{wth_df}{a data frame with columns of weather variables as units vectors}
}
\description{
Drop units from a weather data frame
}
