% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wth_interpolate.R
\name{wth_interpolate}
\alias{wth_interpolate}
\title{Interpolate over missing data}
\usage{
wth_interpolate(x, method = c("linear", "spline"))
}
\arguments{
\item{x}{vector of data to be interpolated}

\item{method}{a character string indicating the method to be used for interpolation}
}
\description{
Interpolate over missing data
}
