% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wth_qc_check.R
\name{wth_qc_check}
\alias{wth_qc_check}
\title{Perform quality controls checks on weather data}
\usage{
wth_qc_check(wth_data, exclude_columns = "DATE", qc_checks = NULL)
}
\arguments{
\item{wth_data}{a data frame containing the weather data to be assessed with
quality control checks}

\item{exclude_columns}{a character string containing the name of the columns
within wth_data to exclude from quality control checks}

\item{qc_checks}{an optional named list of functions that implement
additional quality control checks}
}
\description{
Apply and summarize quality control checks on weather data
}
