% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wth_fill_missing.R
\name{wth_fill_missing}
\alias{wth_fill_missing}
\title{Fill missing data gaps in weather data}
\usage{
wth_fill_missing(
  wth_data,
  fill_data = NULL,
  method = NULL,
  rain_column = "RAIN",
  date_column = "DATE"
)
}
\arguments{
\item{wth_data}{a data frame containing the primary weather data with missing
data gaps to be filled}

\item{fill_data}{an optional data frame containing observations to use for
filling missing data gaps in wth_data}

\item{method}{an optional named list of methods to be used for filling missing
observations where the name of the list element indicates which column and
the value of the list element indicates the method to be used for that column.
Values may include an expression, a numeric constant (e.g. 0), or a
character string indicating the method to be used (i.e. "linear", "spline" or "data").
For any column not listed, the default is to use the matching column from
fill_data ("data"), if provided, assume zero for rainfall (0) and then to
linearly interpolate remaining variables ("linear").}

\item{rain_column}{a character string containing the name of the rain column
within wth_data}

\item{date_column}{a character string containing the name of the date column
within wth_data}
}
\description{
Fill in gaps in observed weather data using a range of techniques including
from a secondary data source, linear interpolation or assumed numeric constant
}
